requirements.txt
Flask==3.0.3
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
joblib==1.4.2
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0